requirements.txt
httpx
pydantic
limits
pillow
exiv2
pathvalidate